requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
rsa==4.9
six==1.16.0
cffi==1.15.1 
cryptography==41.0.3 
ecdsa==0.18.0 
pyasn1==0.5.0 
pycparser==2.21
bcrypt==4.0.1 
passlib==1.7.4